專題競賽/1. 展示程式及資料集/chatbot/requirement.txt
adal==1.2.7
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
archspec @ file:///croot/archspec_1709217642129/work
argcomplete==3.4.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-graphrbac==0.61.1
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==25.4.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azureml-core==1.56.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.3
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1714483155441/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda @ file:///croot/conda_1715635703388/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1714483155348/work
conda_package_streaming @ file:///work/perseverance-python-buildout/croot/conda-package-streaming_1698847176583/work
contextlib2==21.6.0
cryptography @ file:///croot/cryptography_1714660666131/work
distro @ file:///croot/distro_1714488253808/work
dnspython==2.6.1
docker==7.1.0
frozendict @ file:///croot/frozendict_1713194832637/work
h11==0.14.0
humanfriendly==10.0
idna @ file:///croot/idna_1714398848350/work
IMAPClient==3.0.1
isodate==0.6.1
jeepney==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpickle==3.2.2
jsonpointer==2.1
knack==0.11.0
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
menuinst @ file:///croot/menuinst_1714510563922/work
msal==1.29.0
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==2.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
paramiko==3.4.0
pathspec==0.12.1
pkginfo==1.11.1
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
portalocker==2.10.0
pyasn1==0.6.0
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PyNaCl==1.5.0
pyOpenSSL==24.1.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmail36==1.0.5
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==2.0.0
ruamel.yaml==0.16.0
scikit-learn==1.5.0
scipy==1.14.0
SecretStorage==3.3.3
selenium==4.22.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm @ file:///croot/tqdm_1714567712644/work
trio==0.25.1
trio-websocket==0.11.1
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions==4.12.2
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1707770551213/work
websocket-client==1.8.0
wheel==0.43.0
wsproto==1.2.0
zstandard @ file:///croot/zstandard_1714677652653/work
